Refactor ABI-related files into dana/abi
This places the ABI-related files (Encodings.scala) and the
ABI-specific configuration into an `abi` subpackage.
