+++ src/main/scala/dana/Configs.scala
@@ -5,68 +5,12 @@ package dana
 import chisel3._
 import cde._
 import xfiles.{BuildXFilesBackend, XFilesBackendParameters}
+
 import dana.util._
+import dana.abi._
 
-case class GlobalInfo_t (
-  decimal_point: Int,
-  error_function: Int,
-  binary_format: Int,
-  _unused_0: Int,
-  total_weight_blocks: Int,
-  total_neurons: Int,
-  total_layers: Int,
-  ptr_first_layer: Int,
-  ptr_weights: Int
-)
-
-case class LayerInfo_t (
-  ptr_neuron: Int,
-  num_neurons: Int,
-  num_neurons_previous: Int
-)
-
-case class NeuronInfo_t (
-  ptr_weight_offset: Int,
-  num_weights: Int,
-  activation_function: Int,
-  steepness: Int,
-  _unused_0: Int,
-  _unused_1: Int,
-  bias: Int
-)
-
-class DefaultDanaConfig extends Config ( topDefinitions = {
+class DefaultHardwareConfig extends Config ( topDefinitions = {
   (pname,site,here) => pname match {
-    case DanaPtrBits  => 32
-    case DanaDataBits => 32
-
-    case GlobalInfo => GlobalInfo_t (
-      decimal_point       = 3,
-      error_function      = 1,
-      binary_format       = 3,
-      _unused_0           = 9,
-      total_weight_blocks = 16,
-      total_neurons       = 16,
-      total_layers        = 16,
-      ptr_first_layer     = site(DanaPtrBits),
-      ptr_weights         = site(DanaPtrBits))
-
-    case LayerInfo => LayerInfo_t (
-      ptr_neuron           = site(DanaPtrBits),
-      num_neurons          = 16,
-      num_neurons_previous = 16)
-
-    case NeuronInfo => NeuronInfo_t (
-      ptr_weight_offset   = site(DanaPtrBits),
-      num_weights         = 16,
-      activation_function = 5,
-      steepness           = 3,
-      _unused_0           = 8,
-      _unused_1           = 32,
-      bias                = site(DanaDataBits))
-    case DecimalPointOffset => 7
-    case SteepnessOffset => 4
-
     // ANTW Parameters
     case AntwRobEntries            => 32
     // Field widths
@@ -103,9 +47,11 @@ class DefaultDanaConfig extends Config ( topDefinitions = {
   }}
 )
 
+class DefaultDanaConfig extends Config (new Abi32Bit ++
+  new DefaultHardwareConfig)
+
 class DanaNoLearningConfig extends Config ( topDefinitions = {
-  (pname,site,here) =>
-  pname match {
+  (pname,site,here) => pname match {
     case LearningEnabled => false
     case _ => throw new CDEMatchError
   }}
