+++ src/main/scala/dana/Control.scala
@@ -7,6 +7,8 @@ import chisel3._
 import chisel3.util._
 import cde._
 
+import dana.abi._
+
 class ControlCacheInterfaceResp(implicit p: Parameters) extends DanaBundle()(p) {
   val fetch              = Bool()
   val tableIndex         = UInt(log2Up(transactionTableNumEntries).W)
